Fixed preemphasis code totally wrong -- displaying formant bandwidths with gradient

+++ src/analysis/util/calc_formant.cpp
@@ -4,7 +4,7 @@
 Analysis::FormantData Analysis::calculateFormant(float r, float phi, float sampleRate)
 {
     return {
-        .frequency = (float) (fabs(phi) * sampleRate / (2.0f * M_PI)),
-        .bandwidth = (float) (-log(r) * sampleRate / M_PI), 
+        .frequency = (float) (fabsf(phi) * sampleRate / (2.0f * M_PI)),
+        .bandwidth = (float) (-logf(r) * sampleRate / M_PI), 
     };
 }
